+++ CHANGELOG.md
@@ -5,6 +5,7 @@ CHANGELOG
 
 - **New:** A new CLI command which will add class properties to entities similar to stock installation but with slight tweaks (#116)
 - **New:** Prefer template type in title over the type set in URL when creating template (#114)
+- **Fix:** Creating class extension without any options provided throws unique add-on id required error (#115)
 
 ## 1.3.8 (`1030870`)