+++ CHANGELOG.md
@@ -6,6 +6,7 @@ CHANGELOG
 - **New:** A new CLI command which will add class properties to entities similar to stock installation but with slight tweaks (#116)
 - **New:** Prefer template type in title over the type set in URL when creating template (#114)
 - **New:** Display style property group name and display order when viewing style properties (#113)
+- **New:** Allow using full width page in admin control panel (#79)
 - **Fix:** Creating class extension without any options provided throws unique add-on id required error (#115)
 
 ## 1.3.8 (`1030870`)